requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jwt==1.3.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.2
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
sklearn==0.0.post11
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
